Add the annotation in the models to indicate the column Separate the web package in different packages Appropiate annotation for the controller class, add {robotId} in needed mappings Add Annotation SpringApplication in main class Test: change configuration of database (url) and add autowire to the dao etc... Add the annotation in the models to indicate the column Separate the web package in different packages Appropiate annotation for the controller class, add {robotId} in needed mappings Add Annotation SpringApplication in main class Test: change configuration of database (url) and add autowire to the dao etc...
