+++ src/main/java/fr/emse/majeureinfo/springbootintro/SpringBootIntroApplication.java
@@ -1,7 +1,9 @@
 package fr.emse.majeureinfo.springbootintro;
 
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+@SpringBootApplication
 public class SpringBootIntroApplication {
 
 	public static void main(String[] args) {
