+++ src/main/java/fr/emse/majeureinfo/springbootintro/controller/NotFoundException.java
@@ -1,4 +1,4 @@
-package fr.emse.majeureinfo.springbootintro.web;
+package fr.emse.majeureinfo.springbootintro.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
