+++ src/main/java/fr/emse/majeureinfo/springbootintro/model/Actuator.java
@@ -6,10 +6,14 @@ import javax.persistence.*;
 @SuppressWarnings("serial")
 public class Actuator {
 
+    @Id
+    @GeneratedValue
     private Long id;
 
+    @Column
     private Integer speed;
 
+    @Enumerated(EnumType.STRING)
     private Status status;
 
     public Actuator(Integer speed, Status status) {
