+++ src/main/java/fr/emse/majeureinfo/springbootintro/model/Sensor.java
@@ -6,18 +6,20 @@ import javax.persistence.*;
 @SuppressWarnings("serial")
 public class Sensor {
 
+    @Id
+    @GeneratedValue
     private Long id;
 
+    @Column
     private Integer signal;
 
+    @Enumerated(EnumType.STRING)
     private Status status;
 
     public Sensor(Integer signal, Status status) {
         this.signal = signal;
         this.status = status;
     }
-
-
 
     public Long getId() {
         return this.id;
