+++ src/test/resources/test.properties
@@ -2,7 +2,7 @@
 ###
 #   Database Settings
 ###
-spring.datasource.url=jdbc:h2:mem:robots;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.url=jdbc:h2:mem:robotstest;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
 spring.datasource.platform=h2
 spring.datasource.username=sa
 spring.datasource.password=
